Merge browseauctions.jsp with searchauctions.jsp
All of browseauctions.jsp's functionality was included in
searchauctions.jsp's, so I deleted the former and modified the
latter to reflect its added functionality. Users should use
searchauctions.jsp for both browsing and searching purposes.

Index.jsp was updated to reflect these changes. Also removed an
old commented out piece of code that was initially saved for searching
for auctions.

Overall, added documentation and tried to make searchauctions.jsp as
readable as possible for when functionality is later merged in with
other website modules.