Delete duplicate auction.jsp, Add searchauctions updates & bugfixes
Removed the auction.jsp outside of the WebContents folder.

searchauctions.jsp now defaults on Bookname for the item you are
searching for and has a new sort by option called "relevance" which
will perform the expected behavior. Also fixed a bug when searching
for auctions by date.

Small change to userauctions.jsp: now displays the user's username
as opposed to just "user".